Removed parser module
