+++ projects/06/hack_assembler/hack_assembler/tests/test_assembler.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from ..parser import parse_data
+from ..assembler import parse_data
 
 
 def test_parser():
